Refactor: Migrate the assertEquals test to the new BDD format
