+++ unit-tests.lua
@@ -1,4 +1,5 @@
 local testCases = {
+	"./Tests/Builtins/assertions/assertEquals.spec.lua",
 	"./Tests/Extensions/table.spec.lua",
 	"./Tests/Extensions/uv.spec.lua",
 }
